backend/requirements.txt
# aiohappyeyeballs==2.6.1
# aiohttp==3.13.5
# aiosignal==1.4.0
# annotated-doc==0.0.4
# annotated-types==0.7.0
# anyio==4.13.0
# argon2-cffi==25.1.0
# argon2-cffi-bindings==25.1.0
# arrow==1.4.0
# asttokens==3.0.1
# async-lru==2.3.0
# attrs==26.1.0
# babel==2.18.0
# beautifulsoup4==4.14.3
# bleach==6.3.0
# certifi==2026.2.25
# cffi==2.0.0
# charset-normalizer==3.4.6
# click==8.3.1
# colorama==0.4.6
# comm==0.2.3
# dataclasses-json==0.6.7
# debugpy==1.8.20
# decorator==5.2.1
# defusedxml==0.7.1
# distro==1.9.0
# executing==2.2.1
# faiss-cpu==1.13.2
# fastapi==0.135.3
# fastjsonschema==2.21.2
# filelock==3.25.2
# fqdn==1.5.1
# frozenlist==1.8.0
# fsspec==2026.3.0
# greenlet==3.3.2
# groq==0.37.1
# h11==0.16.0
# hf-xet==1.4.3
# httpcore==1.0.9
# httpx==0.28.1
# httpx-sse==0.4.3
# huggingface_hub==1.8.0
# idna==3.11
# ipykernel==7.2.0
# ipython==9.12.0
# ipython_pygments_lexers==1.1.1
# ipywidgets==8.1.8
# isoduration==20.11.0
# jedi==0.19.2
# Jinja2==3.1.6
# jiter==0.13.0
# joblib==1.5.3
# json5==0.14.0
# jsonpatch==1.33
# jsonpointer==3.1.1
# jsonschema==4.26.0
# jsonschema-specifications==2025.9.1
# jupyter==1.1.1
# jupyter-console==6.6.3
# jupyter-events==0.12.0
# jupyter-lsp==2.3.0
# jupyter_client==8.8.0
# jupyter_core==5.9.1
# jupyter_server==2.17.0
# jupyter_server_terminals==0.5.4
# jupyterlab==4.5.6
# jupyterlab_pygments==0.3.0
# jupyterlab_server==2.28.0
# jupyterlab_widgets==3.0.16
# langchain==1.2.14
# langchain-classic==1.0.3
# langchain-community==0.4.1
# langchain-core==1.2.23
# langchain-groq==1.1.2
# langchain-openai==1.1.12
# langchain-text-splitters==1.1.1
# langgraph==1.1.4
# langgraph-checkpoint==4.0.1
# langgraph-prebuilt==1.0.8
# langgraph-sdk==0.3.12
# langsmith==0.7.23
# lark==1.3.1
# markdown-it-py==4.0.0
# MarkupSafe==3.0.3
# marshmallow==3.26.2
# matplotlib-inline==0.2.1
# mdurl==0.1.2
# mistune==3.2.0
# mpmath==1.3.0
# multidict==6.7.1
# mypy_extensions==1.1.0
# nbclient==0.10.4
# nbconvert==7.17.0
# nbformat==5.10.4
# nest-asyncio==1.6.0
# networkx==3.6.1
# notebook==7.5.5
# notebook_shim==0.2.4
# numpy==2.4.4
# openai==2.30.0
# orjson==3.11.7
# ormsgpack==1.12.2
# packaging==26.0
# pandocfilters==1.5.1
# parso==0.8.6
# platformdirs==4.9.4
# prometheus_client==0.24.1
# prompt_toolkit==3.0.52
# propcache==0.4.1
# psutil==7.2.2
# pure_eval==0.2.3
# pycparser==3.0
# pydantic==2.12.5
# pydantic-settings==2.13.1
# pydantic_core==2.41.5
# Pygments==2.20.0
# PyMuPDF==1.27.2.2
# pypdf==6.9.2
# python-dateutil==2.9.0.post0
# python-dotenv==1.2.2
# python-json-logger==4.1.0
# python-multipart==0.0.22
# pywinpty==3.0.3
# PyYAML==6.0.3
# pyzmq==27.1.0
# referencing==0.37.0
# regex==2026.3.32
# requests==2.33.1
# requests-toolbelt==1.0.0
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rfc3987-syntax==1.1.0
# rich==14.3.3
# rpds-py==0.30.0
# safetensors==0.7.0
# scikit-learn==1.8.0
# scipy==1.17.1
# Send2Trash==2.1.0
# sentence-transformers==5.3.0
# setuptools==81.0.0
# shellingham==1.5.4
# six==1.17.0
# sniffio==1.3.1
# soupsieve==2.8.3
# SQLAlchemy==2.0.48
# stack-data==0.6.3
# starlette==1.0.0
# sympy==1.14.0
# tenacity==9.1.4
# terminado==0.18.1
# threadpoolctl==3.6.0
# tiktoken==0.12.0
# tinycss2==1.4.0
# tokenizers==0.22.2
# torch==2.11.0
# tornado==6.5.5
# tqdm==4.67.3
# traitlets==5.14.3
# transformers==5.4.0
# typer==0.24.1
# typing-inspect==0.9.0
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# tzdata==2025.3
# uri-template==1.3.0
# urllib3==2.6.3
# uuid_utils==0.14.1
# uvicorn==0.42.0
# wcwidth==0.6.0
# webcolors==25.10.0
# webencodings==0.5.1
# websocket-client==1.9.0
# widgetsnbextension==4.0.15
# xxhash==3.6.0
# yarl==1.23.0
# zstandard==0.25.0
# actually needed dependencies for the backend, not the entire jupyter ecosystem
fastapi==0.135.3
uvicorn==0.42.0

pydantic==2.12.5
python-dotenv==1.2.2

sqlalchemy==2.0.48
psycopg2-binary==2.9.11

firebase-admin==7.3.0

httpx==0.28.1
requests==2.33.1

beautifulsoup4==4.14.3

langchain==1.2.14
langchain-core==1.2.23
langchain-groq==1.1.2
langgraph==1.1.4
langchain-community==0.4.1

python-multipart==0.0.22

groq==0.37.1

redis==7.4.0